Consolidate two epub lib files
